requirements.txt
ai21==1.0.7
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==3.1.1
anthropic==0.2.9
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cohere==4.5.1
feedparser==6.0.10
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-aiplatform==1.25.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-resource-manager==1.10.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.7
packaging==23.1
Pillow==9.5.0
prettytable==3.7.0
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pyllms==0.2.5
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
tweepy==4.14.0
typing_extensions==4.6.1
urllib3==1.26.16
wcwidth==0.2.6
yarl==1.9.2
